removed trt file, renamed torch module
